+++ storage/memory/log_storage.go
@@ -30,6 +30,7 @@ import (
 	"github.com/google/trillian/monitoring"
 	"github.com/google/trillian/storage"
 	"github.com/google/trillian/storage/cache"
+	stree "github.com/google/trillian/storage/tree"
 	"github.com/google/trillian/types"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -259,6 +260,12 @@ func (t *logTreeTX) WriteRevision(ctx context.Context) (int64, error) {
 		return t.treeTX.writeRevision, errors.New("logTreeTX write revision not populated")
 	}
 	return t.treeTX.writeRevision, nil
+}
+
+// GetMerkleNodes returns the requested nodes at (or below) the read revision.
+func (t *logTreeTX) GetMerkleNodes(ctx context.Context, nodeIDs []stree.NodeID) ([]stree.Node, error) {
+	rev := int64(t.root.Revision)
+	return t.treeTX.subtreeCache.GetNodes(nodeIDs, t.treeTX.getSubtreesAtRev(ctx, rev))
 }
 
 func (t *logTreeTX) DequeueLeaves(ctx context.Context, limit int, cutoffTime time.Time) ([]*trillian.LogLeaf, error) {