+++ storage/mock_storage.go
@@ -501,18 +501,18 @@ func (mr *MockLogTreeTXMockRecorder) GetLeavesByRange(arg0, arg1, arg2 interface
 }
 
 // GetMerkleNodes mocks base method.
-func (m *MockLogTreeTX) GetMerkleNodes(arg0 context.Context, arg1 int64, arg2 []tree.NodeID) ([]tree.Node, error) {
+func (m *MockLogTreeTX) GetMerkleNodes(arg0 context.Context, arg1 []tree.NodeID) ([]tree.Node, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetMerkleNodes", arg0, arg1, arg2)
+	ret := m.ctrl.Call(m, "GetMerkleNodes", arg0, arg1)
 	ret0, _ := ret[0].([]tree.Node)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // GetMerkleNodes indicates an expected call of GetMerkleNodes.
-func (mr *MockLogTreeTXMockRecorder) GetMerkleNodes(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockLogTreeTXMockRecorder) GetMerkleNodes(arg0, arg1 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetMerkleNodes", reflect.TypeOf((*MockLogTreeTX)(nil).GetMerkleNodes), arg0, arg1, arg2)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetMerkleNodes", reflect.TypeOf((*MockLogTreeTX)(nil).GetMerkleNodes), arg0, arg1)
 }
 
 // GetSequencedLeafCount mocks base method.
@@ -946,18 +946,18 @@ func (mr *MockReadOnlyLogTreeTXMockRecorder) GetLeavesByRange(arg0, arg1, arg2 i
 }
 
 // GetMerkleNodes mocks base method.
-func (m *MockReadOnlyLogTreeTX) GetMerkleNodes(arg0 context.Context, arg1 int64, arg2 []tree.NodeID) ([]tree.Node, error) {
+func (m *MockReadOnlyLogTreeTX) GetMerkleNodes(arg0 context.Context, arg1 []tree.NodeID) ([]tree.Node, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetMerkleNodes", arg0, arg1, arg2)
+	ret := m.ctrl.Call(m, "GetMerkleNodes", arg0, arg1)
 	ret0, _ := ret[0].([]tree.Node)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // GetMerkleNodes indicates an expected call of GetMerkleNodes.
-func (mr *MockReadOnlyLogTreeTXMockRecorder) GetMerkleNodes(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockReadOnlyLogTreeTXMockRecorder) GetMerkleNodes(arg0, arg1 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetMerkleNodes", reflect.TypeOf((*MockReadOnlyLogTreeTX)(nil).GetMerkleNodes), arg0, arg1, arg2)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetMerkleNodes", reflect.TypeOf((*MockReadOnlyLogTreeTX)(nil).GetMerkleNodes), arg0, arg1)
 }
 
 // GetSequencedLeafCount mocks base method.