+++ storage/mysql/log_storage.go
@@ -32,6 +32,7 @@ import (
 	"github.com/google/trillian/monitoring"
 	"github.com/google/trillian/storage"
 	"github.com/google/trillian/storage/cache"
+	"github.com/google/trillian/storage/tree"
 	"github.com/google/trillian/types"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -370,6 +371,14 @@ func (t *logTreeTX) WriteRevision(ctx context.Context) (int64, error) {
 		return t.treeTX.writeRevision, errors.New("logTreeTX write revision not populated")
 	}
 	return t.treeTX.writeRevision, nil
+}
+
+// GetMerkleNodes returns the requested nodes at the read revision.
+func (t *logTreeTX) GetMerkleNodes(ctx context.Context, nodeIDs []tree.NodeID) ([]tree.Node, error) {
+	t.treeTX.mu.Lock()
+	defer t.treeTX.mu.Unlock()
+	rev := int64(t.root.Revision)
+	return t.treeTX.subtreeCache.GetNodes(nodeIDs, t.treeTX.getSubtreesAtRev(ctx, rev))
 }
 
 func (t *logTreeTX) DequeueLeaves(ctx context.Context, limit int, cutoffTime time.Time) ([]*trillian.LogLeaf, error) {