+++ CHANGELOG.md
@@ -13,6 +13,11 @@
  * Removed the deprecated crypto.NewSHA256Signer function.
  * Finish removing the `LogMetadata.GetUnsequencedCounts()` method.
 
+### Storage refactoring
+ * `NodeReader.GetMerkleNodes` does not accept revisions anymore. The
+   implementations must use the transaction's `ReadRevision` instead.
+ * TODO(pavelkalinnikov): More changes are coming, and will be added here.
+
 ## v1.3.13
 [Published 2021-02-16](https://github.com/google/trillian/releases/tag/v1.3.13)